\name{checkRank}
\alias{checkRank}
\title{Check k-values manually.}
\usage{
checkRank(ranks, cutoff)
}
\description{Plots a score vs. rank plot with a manually chosen rank cut-off for manual k selection.}
\arguments{
\item{ranks}{getPCRanks output data frame}
\item{cutoff}{integer, rank value to check}
}
