\name{lin}
\alias{lin}
\title{
Find y value of linear regression given x.
}
\description{
Find y value of linear regression given x.
}
\usage{
lin(x, l)
}
\arguments{
\item{x}{x coordinate}
\item{l}{linear model of lm()}
}
\value{
Returns a column \code{numeric} representing the y coordinate at the input \code{x} of the linear model \code{l}.
}
