\name{plot_metagene}
\alias{plot_metagene}
\title{Generate a metagene plot from raw metagene data.}
\usage{
multiple_metagenes(data_list, set_names, title, xaxis, yaxis, legend.title, col, se_alpha)
}
\description{Plots multiple metagene object using the raw data generated by score_metagene().}
\arguments{
\item{data_list}{list}
\item{set_names}{character}
\item{title}{character}
\item{xaxis}{character}
\item{yaxis}{character}
\item{legend.title}{boolean}
\item{col}{character}
\item{se_alpha}{numeric}
}
\value{List of score_metagene() raw data output.}
\value{Character vector of names for score_metagene() object.}
\value{Output plot title}
\value{Output plot x-axis title}
\value{Output plot y-axis title.}
\value{T/F whether to show legend title.}
\value{Vector of colours to use for lines.}
\item{se_alpha}{Alpha value for standard error shading.}
\examples{
multiple_metagenes(list(data_1, data_2, data_3), c("Data 1", "Data 2", "Data 3"))
}
