\name{plot_metagene}
\alias{plot_metagene}
\title{Generate a metagene plot from raw metagene data.}
\usage{
plot_metagene(data, title, xaxis, yaxis, linecol)
}
\description{Plots a metagene object using the raw data generated by score_metagene().}
\arguments{
\item{data}{list}
\item{title}{character}
\item{xaxis}{character}
\item{yaxis}{character}
\item{linecol}{character}
}
\value{score_metagene() raw data output.}
\value{Output plot title.}
\value{Output plot x-axis title}
\value{Output plot y-axis title.}
\value{Colour for line, auto="red".}
\examples{
plot_metagene(metagene_data)
}
