\name{chromDict}
\alias{chromDict}
\title{Convert a rank object into a chromDict.}
\usage{
chromDict(mat)
}
\description{
Internal to many functions; creates a chromDict for faster computing times. chromDict can be run separately to speed up functions run iteratively.}
\arguments{
\item{ranks}{getPCRanks output data frame}
}
