\name{multiple_metagenes}
\alias{multiple_metagenes}
\title{Plot multiple metagene data objects on a single plot.}
\usage{
multiple_metagenes(data_list, set_names, title, xaxis, yaxis, legend.title, col, se_alpha)
}
\description{Plots multiple metagene object using the raw data generated by score_metagene().}
\arguments{
\item{data_list}{List of score_metagene() raw data output}
\item{set_names}{Character vector of names for score_metagene() object}
\item{title}{Output plot title}
\item{xaxis}{Output plot x-axis title}
\item{yaxis}{Output plot y-axis title}
\item{legend.title}{T/F, whether to show legend title}
\item{col}{Vector of colours to use for lines}
\item{se_alpha}{0-1, alpha value for standard error shading}
}

