\name{plot_metagene}
\alias{plot_metagene}
\title{Generate a metagene plot from raw metagene data.}
\usage{
plot_metagene(data, title, xaxis, yaxis, linecol)
}
\description{Plots a metagene object using the raw data generated by score_metagene().}
\arguments{
\item{data}{list, score_metagene() raw data output}
\item{title}{Output plot title}
\item{xaxis}{Output plot x-axis title}
\item{yaxis}{Output plot y-axis title}
\item{linecol}{Colour for line, auto="red"}
}
